requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
cached-property==1.5.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chroma-hnswlib==0.7.2
chromadb==0.4.5
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
cryptography==41.0.2
dataclasses-json==0.5.13
et-xmlfile==1.1.0
excel2img==1.4.0
exceptiongroup==1.1.2
fastapi==0.99.1
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
google-search-results==2.4.2
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
isodate==0.6.1
joblib==1.3.1
langchain==0.0.240
langsmith==0.0.14
lxml==4.9.2
Markdown==3.4.3
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
platformdirs==3.6.0
posthog==3.0.1
protobuf==4.24.0
pulsar-client==3.2.0
pycparser==2.21
pydantic==1.10.11
pymongo==4.1.1
pypandoc==1.11
PyPika==0.48.9
pypiwin32==223
pyreadline3==3.4.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.13.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pywin32==306
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.8.1
urllib3==2.0.3
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
xlrd==2.0.1
XlsxWriter==3.0.3
yarl==1.9.2
zeep==4.1.0
